Fix CI failures for style-object fixtures
- Split style-object-static and style-object-dynamic into separate fixture
  files (generate-expected-html script expects fixture.id.ts file paths)
- Add style-object-static to CSR skip list (stateless component; attribute
  ordering differs between SSR and CSR injection)
- Add styleToCss to CSR mock runtime so dynamic style fixtures can evaluate
- Add style-object-dynamic to Mojolicious skip list (Perl templates cannot
  handle JS object literal expressions, same as Go templates)
