+++ packages/adapter-tests/fixtures/index.ts
@@ -28,7 +28,9 @@ import { fixture as mapDynamicClass } from './map-dynamic-class'
 import { fixture as voidElements } from './void-elements'
 import { fixture as dynamicAttributes } from './dynamic-attributes'
 import { fixture as classVsClassname } from './class-vs-classname'
-import { fixture as styleAttribute, fixtureStaticObject as styleObjectStatic, fixtureDynamicObject as styleObjectDynamic } from './style-attribute'
+import { fixture as styleAttribute } from './style-attribute'
+import { fixture as styleObjectStatic } from './style-object-static'
+import { fixture as styleObjectDynamic } from './style-object-dynamic'
 // Priority 6: Advanced patterns
 import { fixture as fragment } from './fragment'
 import { fixture as fragmentConditional } from './fragment-conditional'
